core/requirements.txt
marshmallow==3.19.0
requests==2.28.2
tornado==6.3.1
pymongo==4.3.3
cryptography==40.0.2
python-jose==3.3.0
python-keycloak==2.16.3
Pygments==2.15.1
influxdb-client==1.36.1
PyYAML==6.0
